requirements_databricks.txt
databricks-sdk>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
